moved tenant rebalance related classes to a separate package
