+++ pinot-controller/src/main/java/org/apache/pinot/controller/helix/core/rebalance/tenant/TenantRebalancer.java
@@ -16,23 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.controller.helix.core.rebalance;
+package org.apache.pinot.controller.helix.core.rebalance.tenant;
 
-public interface TenantRebalanceObserver {
-  enum Trigger {
-    // Start of tenant rebalance Trigger
-    START_TRIGGER,
-    // rebalance of a table is started
-    REBALANCE_STARTED_TRIGGER,
-    // rebalance of a table is completed
-    REBALANCE_COMPLETED_TRIGGER,
-    // rebalance of a table is failed
-    REBALANCE_ERRORED_TRIGGER
-  }
 
-  void onTrigger(TenantRebalanceObserver.Trigger trigger, String tableName, String description);
-
-  void onSuccess(String msg);
-
-  void onError(String errorMsg);
+public interface TenantRebalancer {
+  TenantRebalanceResult rebalance(TenantRebalanceContext context);
 }
